Enum linera_alloy_rpc_client::BuiltInConnectionString

#[non_exhaustive]
pub enum BuiltInConnectionString {
    Http(Url),
}

Connection string for built-in transports.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Http(Url)

HTTP transport.

Implementations

impl BuiltInConnectionString

pub async fn connect_boxed(&self) -> Result<BoxTransport, TransportError>

Connect with the given connection string.

Notes

If hyper feature is enabled

WS will extract auth, however, auth is disabled for wasm.

pub fn try_as_http(s: &str) -> Result<Self, TransportError>

Tries to parse the given string as an HTTP URL.