Struct linera_alloy_rpc_client::PollChannel

pub struct PollChannel<Resp> { /* private fields */ }

A channel yielding responses from a poller task.

This stream is backed by a coroutine, and will continue to produce responses until the poller task is dropped. The poller task is dropped when all RpcClient instances are dropped, or when all listening PollChannel are dropped.

The poller task also ignores errors from the server and deserialization errors, and will continue to poll until the client is dropped.

Implementations

impl<Resp> PollChannel<Resp>
where Resp: RpcReturn + Clone,

pub fn resubscribe(&self) -> Self

Resubscribe to the poller task.

pub fn into_stream(self) -> impl Stream<Item = Resp> + Unpin

Converts the poll channel into a stream.

pub fn into_stream_raw(self) -> BroadcastStream<Resp>

Converts the poll channel into a stream that also yields lag errors.