Struct linera_alloy_rpc_client::RpcCall

pub struct RpcCall<Conn, Params, Resp, Output = Resp, Map = fn(_: Resp) -> Output>
where
    Conn: Transport + Clone,
    Params: RpcParam,
    Map: Fn(Resp) -> Output,
{ /* private fields */ }

A prepared, but unsent, RPC call.

This is a future that will send the request when polled. It contains a Request, a Transport, and knowledge of its expected response type. Upon awaiting, it will send the request and wait for the response. It will then deserialize the response into the expected type.

Errors are captured in the RpcResult type. Rpc Calls will result in either a successful response of the Resp type, an error response, or a transport error.

Note

Serializing the request is done lazily. The request is not serialized until the future is polled. This differs from the behavior of crate::BatchRequest, which serializes greedily. This is because the batch request must immediately erase the Param type to allow batching of requests with different Param types, while the RpcCall may do so lazily.

Implementations

impl<Conn, Params, Resp, Output, Map> RpcCall<Conn, Params, Resp, Output, Map>
where Conn: Transport + Clone, Params: RpcParam, Map: Fn(Resp) -> Output,

pub fn map_resp<NewOutput, NewMap>( self, map: NewMap, ) -> RpcCall<Conn, Params, Resp, NewOutput, NewMap>
where NewMap: Fn(Resp) -> NewOutput,

Set a function to map the response into a different type.

pub fn is_subscription(&self) -> bool

Returns true if the request is a subscription.

Panics

Panics if called after the request has been sent.

pub fn set_is_subscription(&mut self)

Set the request to be a non-standard subscription (i.e. not โeth_subscribeโ).

Panics

Panics if called after the request has been sent.

pub fn set_subscription_status(&mut self, status: bool)

Set the subscription status of the request.

pub fn params(&mut self) -> &mut Params

Get a mutable reference to the params of the request.

This is useful for modifying the params after the request has been prepared.

Panics

Panics if called after the request has been sent.

pub fn request(&self) -> &Request<Params>

Returns a reference to the request.

Panics

Panics if called after the request has been sent.

pub fn request_mut(&mut self) -> &mut Request<Params>

Returns a mutable reference to the request.

Panics

Panics if called after the request has been sent.

impl<Conn, Params, Resp, Output, Map> RpcCall<Conn, &Params, Resp, Output, Map>
where Conn: Transport + Clone, Params: RpcParam + Clone, Map: Fn(Resp) -> Output,

pub fn into_owned_params(self) -> RpcCall<Conn, Params, Resp, Output, Map>

Convert this call into one with owned params, by cloning the params.

Panics

Panics if called after the request has been sent.

impl<'a, Conn, Params, Resp, Output, Map> RpcCall<Conn, Params, Resp, Output, Map>
where Conn: Transport + Clone, Params: RpcParam + 'a, Resp: RpcReturn, Output: 'static, Map: Fn(Resp) -> Output + Send + 'a,

pub fn boxed(self) -> RpcFut<'a, Output>

Convert this future into a boxed, pinned future, erasing its type.