Struct linera_alloy_rpc_client::RpcClient

pub struct RpcClient<T>(/* private fields */);

A JSON-RPC client.

RpcClient should never be instantiated directly. Instead, use ClientBuilder.

Implementations

impl RpcClient<Identity>

pub fn builder() -> ClientBuilder<Identity>

Create a new ClientBuilder.

impl RpcClient<Http<Client>>

pub fn new_http(url: Url) -> Self

Create a new RpcClient with an HTTP transport.

impl<T> RpcClient<T>

pub fn new(t: T, is_local: bool) -> Self

Creates a new RpcClient with the given transport.

pub fn from_inner(inner: RpcClientInner<T>) -> Self

Creates a new RpcClient with the given inner client.

pub const fn inner(&self) -> &Arc<RpcClientInner<T>>

Get a reference to the client.

pub fn into_inner(self) -> Arc<RpcClientInner<T>>

Convert the client into its inner type.

pub fn get_weak(&self) -> WeakClient<T>

Get a Weak reference to the client.

pub fn get_ref(&self) -> ClientRef<'_, T>

Borrow the client.

pub fn with_poll_interval(self, poll_interval: Duration) -> Self

Sets the poll interval for the client in milliseconds.

Note: This will only set the poll interval for the client if it is the only reference to the inner client. If the reference is held by many, then it will not update the poll interval.

impl<T: Transport> RpcClient<T>

pub fn prepare_static_poller<Params, Resp>( &self, method: impl Into<Cow<'static, str>>, params: Params, ) -> PollerBuilder<T, Params, Resp>
where T: Clone, Params: RpcParam + 'static, Resp: RpcReturn + Clone,

Build a poller that polls a method with the given parameters.

See PollerBuilder for examples and more details.

impl<T: Transport + Clone> RpcClient<T>

pub fn boxed(self) -> RpcClient<BoxTransport>

Boxes the transport.

This will create a new client if this instance is not the only reference to the inner client.

impl<T> RpcClient<Http<T>>

pub fn new_batch(&self) -> BatchRequest<'_, Http<T>>

Create a new BatchRequest builder.